Qt: extend drag and drop to mainwindow
